add test job
Run tests before container release, if there are any problems the job fails and the images are not released. At the moment this is basic, but it should cover the most serious issues that may appear.

This will prevent fiascos like https://github.com/searxng/searxng/issues/4718 to ever happen again (hopefully)